requirements.txt
ultralytics>=8.1.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
dvc[s3]>=3.0.0
dvclive>=3.0.0
mlflow>=2.12.0
pypdfium2>=4.30.0
pillow>=10.0.0
pyyaml>=6.0.1
pandas>=2.1.0
tqdm>=4.66.0
rfdetr>=0.1.0

